+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * `chunk_texts()` now outputs texts that keep the same spaces present in the original and no longer outputs texts with spaces around the punctuation marks.
 
+* `chunk_texts()` now also accepts tokens objects as input and in that case returns chunks of sentences for which the total length is equal or greater than the one specified.
+
 # idiolect 1.2.0
 
 * minor bug fixes